Initial commit for rewrite
Starting strong with syntax checks!
