+++ Prebuild/src/Core/Platform.h
@@ -1,6 +1,8 @@
 #pragma once
 #include "Core.h"
 
+#include <string>
+
 namespace Prebuild
 {
     class Platform
@@ -11,28 +13,44 @@ namespace Prebuild
 
         static Scope<Platform> Create(const char* platformType);
 
+        const size_t NPOS = std::string::npos;
+
         enum KindType
         {
-            NONE = 0,
-            STATICLIB = 1,
-            DYNAMICLIB = 2,
+            kNONE      = 0,
+            STATICLIB  = 1,
+            SHAREDLIB  = 2,
             CONSOLEAPP = 3,
+        };
+
+        enum LanguageType
+        {
+            lNONE = 0,
+            C     = 1,
+            CXX   = 2,
+        };
+
+        enum ArchitectureType
+        {
+            aNONE = 0,
+            X86 = 1,
+            X64 = 2,
         };
 
         const char* Keywords[10] = 
         {
             // WORKSPACE
             "workspace",
-            "version",
             "architecture",
-            "flags",
+            "configurations",
+            "defines",
             // PROJECTS
             "project",
-            "mainfile",
             "kind",
             "files",
             "includedirs",
-            "links"
+            "links",
+            "configurations",
         };
 
         const char* PathKeywords[1] =
